FRO-433: Fix: Resolve responsive issues and refactor code in Quick Mode (settings, review, and questionnaire pages)
