+++ src/components/dashboard/dashbordContainer.tsx
@@ -33,33 +33,6 @@ import { ArrowForward } from "@mui/icons-material";
 import { useAssessmentContext } from "@/providers/AssessmentProvider";
 import { ASSESSMENT_MODE } from "@/utils/enumType";
 
-// ---------- Types
-type TabItem = {
-  label: string;
-  address: string;
-  permission: string;
-};
-
-const tabListTitle: TabItem[] = [
-  { label: "dashboard", address: "dashboard", permission: "viewDashboard" },
-  {
-    label: "questions",
-    address: "questionnaires",
-    permission: "viewAssessmentQuestionnaireList",
-  },
-  {
-    label: "insights",
-    address: "insights",
-    permission: "viewAssessmentInsights",
-  },
-  { label: "advice", address: "advice", permission: "createAdvice" },
-  {
-    label: "reportTitle",
-    address: "report",
-    permission: "manageReportMetadata",
-  },
-];
-
 const maxLength = 40;
 
 const DashbordContainer: React.FC = () => {
@@ -69,16 +42,10 @@ const DashbordContainer: React.FC = () => {
   const { assessmentId = "" } = useParams<{ assessmentId?: string }>();
   const outlet = useOutlet();
   const buttonRef = useRef<HTMLButtonElement | null>(null);
-  const { assessmentInfo } = useAssessmentContext();
+  const { assessmentInfo, permissions } = useAssessmentContext();
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
 
   const open = Boolean(anchorEl);
-  const handleClick = (event: React.MouseEvent<HTMLElement>) =>
-    setAnchorEl(event.currentTarget);
-  const handleClose = () => setAnchorEl(null);
-
-  const showTabName = () =>
-    t(tabListTitle.find((item) => item.address === selectedTab)?.label ?? "");
 
   useEffect(() => {
     const pathSegments = location.pathname.split("/").filter(Boolean);
@@ -94,7 +61,6 @@ const DashbordContainer: React.FC = () => {
     runOnMount: true,
   });
 
-  const isMobileScreen = useScreenResize("sm");
   const titleLength = assessmentInfo?.title?.length ?? 0;
 
   return (
@@ -103,7 +69,11 @@ const DashbordContainer: React.FC = () => {
       renderLoading={() => <LoadingSkeletonOfAssessmentRoles />}
       render={([pathInfo]) => (
         <Box sx={{ ...styles.centerCV }} m="auto" pb={3} gap={1}>
-          <DashboardTitle pathInfo={pathInfo} title={assessmentInfo?.title} />
+          <DashboardTitle
+            pathInfo={pathInfo}
+            title={assessmentInfo?.title}
+            permissions={permissions}
+          />
           <Grid
             container
             columns={12}
@@ -147,12 +117,14 @@ const DashbordContainer: React.FC = () => {
                   textAlign="left"
                   variant="headlineLarge"
                   sx={{
-                    fontFamily: languageDetector(assessmentInfo?.title)
+                    fontFamily: languageDetector(
+                      pathInfo?.assessment?.title ?? assessmentInfo?.title,
+                    )
                       ? farsiFontFamily
                       : primaryFontFamily,
                   }}
                 >
-                  {assessmentInfo?.title}
+                  {pathInfo?.assessment?.title ?? assessmentInfo?.title}
                 </Typography>
               )}
             </Grid>
@@ -166,76 +138,11 @@ const DashbordContainer: React.FC = () => {
               sx={{ display: "flex", my: 1 }}
               justifyContent="flex-end"
             >
-              {isMobileScreen ? (
-                <Box
-                  sx={{
-                    ...styles.centerVH,
-                    mt: 1,
-                    background: "#2466A814",
-                    borderRadius: "16px",
-                    p: 1.3,
-                  }}
-                >
-                  <Button
-                    ref={buttonRef}
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      handleClick(e);
-                    }}
-                    endIcon={
-                      open ? (
-                        <ArrowDropUpRoundedIcon />
-                      ) : (
-                        <ArrowDropDownRoundedIcon />
-                      )
-                    }
-                  >
-                    {showTabName()}
-                  </Button>
-                  <Menu
-                    id="basic-menu"
-                    anchorEl={anchorEl}
-                    open={open}
-                    onClose={handleClose}
-                    PaperProps={{
-                      sx: {
-                        width: buttonRef.current
-                          ? `${buttonRef.current.offsetWidth}px`
-                          : "180px",
-                      },
-                    }}
-                  >
-                    {tabListTitle.map(({ label, address }) => (
-                      <MenuItem
-                        key={label}
-                        dense
-                        component={NavLink}
-                        to={address}
-                        onClick={handleClose}
-                        sx={{
-                          fontFamily: languageDetector(label)
-                            ? farsiFontFamily
-                            : primaryFontFamily,
-                          ...theme.typography.semiBoldMedium,
-                          fontSize: "14px",
-                          transition: "all 0.2s ease",
-                        }}
-                      >
-                        <ListItemText>
-                          <Trans i18nKey={label} />
-                        </ListItemText>
-                      </MenuItem>
-                    ))}
-                  </Menu>
-                </Box>
-              ) : (
-                <MainTabs
-                  onTabChange={handleTabChange}
-                  selectedTab={selectedTab}
-                  tabListTitle={tabListTitle}
-                  flexColumn={titleLength < maxLength}
-                />
-              )}
+              <MainTabs
+                onTabChange={handleTabChange}
+                selectedTab={selectedTab}
+                flexColumn={titleLength < maxLength}
+              />
             </Grid>
             {/* Outlet */}
             <Grid container mt={2}>
